+++ cmd/tools/gml2gv.h
@@ -5,40 +5,53 @@
 
 #include <stdio.h>
 #include <cgraph/cgraph.h>
+#include <cgraph/list.h>
 
 typedef struct {
-    Dtlink_t link;
     unsigned short kind;
     unsigned short sort;
     char* name;
     union {
 	char* value;
-	Dt_t* lp;
+	void *lp; ///< actually an `attrs_t *`
     }u;
 } gmlattr;
 
+void free_attr(gmlattr *p);
+
+DEFINE_LIST_WITH_DTOR(attrs, gmlattr *, free_attr)
+
 typedef struct {
-    Dtlink_t link;
     char* id;
-    Dt_t* attrlist;  
+    attrs_t attrlist;  
 } gmlnode;
 
+void free_node(gmlnode *p);
+
+DEFINE_LIST_WITH_DTOR(nodes, gmlnode *, free_node)
+
 typedef struct {
-    Dtlink_t link;
     char* source;
     char* target;
-    Dt_t* attrlist;  
+    attrs_t attrlist;  
 } gmledge;
 
+void free_edge(gmledge *p);
+
+DEFINE_LIST_WITH_DTOR(edges, gmledge *, free_edge)
+
 typedef struct gmlgraph {
-    Dtlink_t link;
     struct gmlgraph* parent;
     int directed;
-    Dt_t* attrlist;  
-    Dt_t* nodelist;  
-    Dt_t* edgelist;  
-    Dt_t* graphlist;  
+    attrs_t attrlist;  
+    nodes_t nodelist;
+    edges_t edgelist;
+    void *graphlist; ///< actually a `graphs_t *`
 } gmlgraph;
+
+void free_graph(gmlgraph *p);
+
+DEFINE_LIST_WITH_DTOR(graphs, gmlgraph *, free_graph)
 
 extern int gmllex(void);
 extern void gmllexeof(void);