+++ lib/cdt/dtflatten.c
@@ -12,12 +12,12 @@ Dtlink_t* dtflatten(Dt_t* dt)
 	Dtlink_t	*t, *r, *list, *last, **s, **ends;
 
 	/* already flattened */
-	if(dt->data->type&DT_FLATTEN )
-		return dt->data->here;
+	if (dt->data.type & DT_FLATTEN )
+		return dt->data.here;
 
 	list = last = NULL;
-	if(dt->data->type&DT_SET)
-	{	for(ends = (s = dt->data->htab) + dt->data->ntab; s < ends; ++s)
+	if (dt->data.type & DT_SET)
+	{	for (ends = (s = dt->data.htab) + dt->data.ntab; s < ends; ++s)
 		{	if((t = *s) )
 			{	if(last)
 					last->right = t;
@@ -28,9 +28,7 @@ Dtlink_t* dtflatten(Dt_t* dt)
 			}
 		}
 	}
-	else if(dt->data->type&DT_QUEUE)
-		list = dt->data->head;
-	else if((r = dt->data->here) ) /*if(dt->data->type&(DT_OSET|DT_OBAG))*/
+	else if ((r = dt->data.here)) // if (dt->data.type & (DT_OSET|DT_OBAG))
 	{	while((t = r->left) )
 			RROTATE(r,t);
 		for(list = last = r, r = r->right; r; last = r, r = r->right)
@@ -43,8 +41,8 @@ Dtlink_t* dtflatten(Dt_t* dt)
 		}
 	}
 
-	dt->data->here = list;
-	dt->data->type |= DT_FLATTEN;
+	dt->data.here = list;
+	dt->data.type |= DT_FLATTEN;
 
 	return list;
 }