+++ lib/cdt/dthdr.h
@@ -24,8 +24,7 @@
 #define HLOAD(s)	((s) << 1)
 #define HINDEX(n,h)	((h)&((n)-1))
 
-#define UNFLATTEN(dt) \
-		((dt->data->type&DT_FLATTEN) ? dtrestore(dt,NULL) : 0)
+#define UNFLATTEN(dt) ((dt->data.type & DT_FLATTEN) ? dtrestore(dt, NULL) : 0)
 
 /* tree rotation/linking functions */
 #define rrotate(x,y)	((x)->left  = (y)->right, (y)->right = (x))