+++ lib/cdt/dtmethod.c
@@ -18,34 +18,21 @@ Dtmethod_t* dtmethod(Dt_t* dt, Dtmethod_t* meth)
 	/* get the list of elements */
 	list = dtflatten(dt);
 
-	if(dt->data->type&DT_QUEUE)
-		dt->data->head = NULL;
-	else if(dt->data->type&DT_SET)
-	{	if(dt->data->ntab > 0)
-			free(dt->data->htab);
-		dt->data->ntab = 0;
-		dt->data->htab = NULL;
+	if (dt->data.type & DT_SET)
+	{	if (dt->data.ntab > 0)
+			free(dt->data.htab);
+		dt->data.ntab = 0;
+		dt->data.htab = NULL;
 	}
 
-	dt->data->here = NULL;
-	dt->data->type = (dt->data->type&~(DT_METHODS|DT_FLATTEN)) | meth->type;
+	dt->data.here = NULL;
+	dt->data.type = (dt->data.type & ~(DT_METHODS|DT_FLATTEN)) | meth->type;
 	dt->meth = meth;
 	if(dt->searchf == oldmeth->searchf)
 		dt->searchf = meth->searchf;
 
-	if(meth->type&DT_QUEUE)
-	{	if(!(oldmeth->type&DT_QUEUE))
-		{	if((r = list) )
-			{	Dtlink_t*	t;
-				for(t = r->right; t; r = t, t = t->right )
-					t->left = r;
-				list->left = r;
-			}
-		}
-		dt->data->head = list;
-	}
-	else if(meth->type&(DT_OSET|DT_OBAG))
-	{	dt->data->size = 0;
+	if(meth->type&(DT_OSET|DT_OBAG))
+	{	dt->data.size = 0;
 		while(list)
 		{	r = list->right;
 			meth->searchf(dt, list, DT_RENEW);
@@ -58,7 +45,8 @@ Dtmethod_t* dtmethod(Dt_t* dt, Dtmethod_t* meth)
 			rehash = 1;
 		else	rehash = 0;
 
-		dt->data->size = dt->data->loop = 0;
+		dt->data.size = 0;
+		dt->data.loop = 0;
 		while(list)
 		{	r = list->right;
 			if(rehash)