+++ CHANGELOG.md
@@ -27,6 +27,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: The cgraph macros `agnew` and `agnnew` have been removed.
 - **Breaking**: The cgraph functions `agalloc`, `agrealloc`, and `agfree` have
   been removed.
+- **Breaking**: The cdt queue types, `DT_QUEUE`, `Dtqueue`, `_Dtqueue` have been
+  removed.
+- **Breaking**: The cdt aliases `struct _dthold_s`, `struct _dtdata_s`,
+  `struct _dtmethod_s`, `struct _dtstat_s` have been removed.
+- **Breaking**: The cdt type `_Dttree` has been removed.
+- **Breaking**: The cdt aliases `struct _dtlink_s`, `struct _dtdisc_s`, and
+  `struct _dt_s` have been renamed to `struct dtlink_s_`, `struct dtdisc_s_`,
+  and `struct dt_s_` respectively.
+- **Breaking**: The cdt macro `DT_PRIME` has been removed.
+- **Breaking**: The cdt field `Dt_t.data` is now a `Dtdata_t` instead of
+  `Dtdata_t *`.
 
 ### Fixed